run-tests: check PR author instead of branch owner
We only run tests when the branch owner has write permissions to the
repository under test.

Some people seem to prefer opening pull requests from a branch in the
upstream repository instead of their fork. This made our test fail,
because the branch owner is the organization in that case and GitHub
doesn't seem to have permissions for organizations.

Check the author of the pull request instead to support this use case.